pull out class methods, getters and setters, missing dates functions
Pulled out class methods, not sure if it's a good idea since they are meant to only be used with the class... but meh

getters and setters for possible structural changes in the future

functions to know if there are missing dates and also to get the  ranges of missing dates.

dindex updated to be able to round up or down the input date within the dates array.