get set iterators and scalabilty improvements
Made the getters able to take an array of indices or an iterator.

add static decorators.

used new getters to fix a couple of static methods using the instance variables.

Fixed dindex to round up or down and added a round_date function that works similarly and returns the date.