Finally split everything into pkgs!
Manage to split everything into packages, I'll add a diagram of what I'm thinking but the directory structure should reflect it somehow.

Also made a small fix to the get time window to return an empty array if the query is out of range to have a consistent default behaviour

I also made some changes in the logic of some plotting functions, I still need to work on it but it is now super easy to plot from a WL element!