Introduced frequencies and related functions
I added the delta function used to know how big is a time step is between a date and the next in the time series

added a function to normalize a date, i.e. fit it into the format of the TS object, so if monthly then 14 of feb becomes 1 of feb, etc.